curator/requirements.txt
paho_mqtt==1.6.1
geopy==2.3.0
numpy==1.24.4
pandas==2.0.0
python_dateutil==2.8.2
pytz==2021.3
requests==2.31.0
